refactor: consolidate bridge endpoints onto a shared labels-view
Both /api/public/is-known and /api/public/metadata pipelines previously
fetched products.json + entities.json + earn-vaults.json + escrow
perspective + snapshot independently and built their own productByVault
and entityAddresses maps. The two pipelines were structurally parallel
but mechanically independent — exactly the drift surface the shared
governor rule was meant to prevent.

New server/utils/labels-view.ts assembles one per-chain view with
snapshot + all label-derived maps + verificationLabels. Both endpoints
consume it. In-flight dedup at the view level collapses concurrent
rebuilds (e.g. the two warm-cache tasks fire in parallel) onto a single
upstream pass.

Net: vault-metadata 535→303, verified-vaults 141→82, labels-helpers
109→51, new labels-view 282 — saved ~67 lines and centralized the
substrate. Dropped buildProductMaps and buildDeprecatedEarnSet from
labels-helpers (they were used by tests only after vault-metadata
gained its own richer ProductDescriptor).
